image_convert/requirements.txt
python-swiftclient==3.10.1
pillow==7.2.0
